+++ Assets/FluentMotion/IReactiveObject.cs
@@ -3,12 +3,10 @@ using UniRx;
 
 namespace FluentMotion
 {
-    public interface IReactiveObject<in T, U>
+    public interface IReactiveObject<in T, out U>
     {
         U Selector(T reactiveObject);
-        Subject<U> Subject { get; }
         bool IsValid(T reactiveObject);
-
         IObservable<U> AsObservable();
     }
 }