+++ Assets/FluentMotion/finger/impl/Thumb.cs
@@ -6,5 +6,6 @@ namespace FluentMotion.finger.impl
     public static class Thumb
     {
         public static bool IsExtended(Hand hand) => hand.GetThumb().IsExtended;
+        public static bool IsExtended(Finger finger) => finger.IsExtended;
     }
 }